requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-xai
pypdf
chromadb
python-dotenv
sentence-transformers
langchain-groq